split masker and label masker
