+++ nilearn/maskers/__init__.py
@@ -8,7 +8,8 @@ from .nifti_labels_masker import NiftiLabelsMasker
 from .nifti_maps_masker import NiftiMapsMasker
 from .nifti_masker import NiftiMasker
 from .nifti_spheres_masker import NiftiSpheresMasker
-from .surface_maskers import SurfaceLabelsMasker, SurfaceMasker
+from .surface_labels_masker import SurfaceLabelsMasker
+from .surface_maskers import SurfaceMasker
 
 __all__ = [
     "BaseMasker",
